BOX-112: add skeleton for user page, also combine author page and user page
