+++ src/pages/paths.ts
@@ -3,8 +3,8 @@ export const paths = {
   /** @test */
   comments: () => '/comments',
   card: (cardId: string) => `/card/${cardId}`,
-  user: () => '/user',
-  author: (username: string) => `/user/${username}`,
+  // user: () => '/user',
+  user: (username: string) => `/u/${username}`,
   search: (query = '') => {
     if (query) return `/search?query=${query}`;
     return '/search';
